GInteraction, can use think in slash command
